scripts/requirements.txt
MLB-StatsAPI==1.8.1
boto3==1.39.2